pyproject.toml
[tool.poetry]
name = "goit-pyweb-hw-14"
version = "0.1.0"
description = ""
authors = ["Sergii Ternytskyi <serhij86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"
alembic = "^1.13.2"
uvicorn = {extras = ["standard"], version = "^0.30.3"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-jwt-auth = "^0.5.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
fastapi-mail = "^1.4.1"
libgravatar = "^1.0.4"
redis = "^5.0.8"
cloudinary = "^1.41.0"
python-dotenv = "^1.0.1"
fastapi-limiter = "^0.1.6"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.0.2"


[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
